Still smarter landing behavior: Planes take into account other planes already en route to targets so they don't have wait for the other planes to actually land to find a new landing spot later and make a detour.
If no landing spot in range is free, other planes with longer lasting
fuel supplies are rerouted; bombers with nukes are priorised, of course.

If that fails, too, planes head towards the nearest not horribly
overbooked landing pad and cruise about, hoping to get noticed.
In previous versions, they'd head for the closest pad with open spots.